requirements.txt
fastapi==0.104.1
pydantic==2.9.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
Pillow==11.1.0
google-cloud-vision==3.7.2

